Decode error bitmask (#20)

+++ components/daly_bms_ble/daly_bms_ble.h
@@ -146,7 +146,7 @@ class DalyBmsBle : public esphome::ble_client::BLEClientNode, public PollingComp
   void publish_state_(sensor::Sensor *sensor, float value);
   void publish_state_(switch_::Switch *obj, const bool &state);
   void publish_state_(text_sensor::TextSensor *text_sensor, const std::string &state);
-  std::string bitmask_to_string_(const char *const messages[], const uint8_t &messages_size, const uint8_t &mask);
+  std::string bitmask_to_string_(const char *const messages[], const uint8_t &messages_size, const uint64_t &mask);
 
   bool check_bit_(uint16_t mask, uint16_t flag) { return (mask & flag) == flag; }
 };
